# docs/openapi.yaml
openapi: 3.1.0
info:
  title: "HelioBio-Social API"
  description: "Sistema de Análisis Heliobiológico con Alertas Inteligentes - Versión 1.3.0"
  version: "1.3.0"
  contact:
    name: "HelioBio-Social Research Team"
    email: "research@heliobio-social.org"
  license:
    name: "MIT"
    url: "https://opensource.org/licenses/MIT"
servers:
  - url: http://localhost:8000
    description: Servidor de desarrollo local
  - url: https://api.heliobio-social.org/v1
    description: Servidor de producción
paths:
...
  /api/alerts/{alert_id}/acknowledge:
    post:
      summary: "Reconocer Alerta"
      description: "Marca una alerta como reconocida por el usuario"
      parameters:
        - name: alert_id
          in: path
          required: true
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: "Alerta reconocida correctamente"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "acknowledged"
                  alert_id:
                    type: integer
                    example: 1
        '404':
          description: "Alerta no encontrada"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/historical/data:
    get:
      summary: "Obtener Datos Históricos"
      description: "Retorna datos históricos para análisis temporal"
      parameters:
        - name: hours
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 168
            default: 6
          description: "Número de horas de datos históricos a retornar"
      responses:
        '200':
          description: "Datos históricos obtenidos correctamente"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoricalResponse'
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT token para autenticación (futura implementación)"
security:
  - BearerAuth: []